requirements.txt
Flask==2.3.3
pymongo[srv]==4.6.1
google-cloud-texttospeech==2.14.1
gunicorn==21.2.0
dnspython==2.4.2 